Refactor concepts code into new panel
